changed svg properties to camelCase
