--- assets/Icons/icons.tsx
@@ -1,199 +1,0 @@
-export const CheckmarkIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="8"
-    height="7"
-    viewBox="0 0 8 7"
-    fill="none"
-  >
-    <path
-      d="M7.66153 6.5C7.66153 6.77614 7.43768 7 7.16153 7H1.26923C0.993084 7 0.769226 6.77614 0.769226 6.5C0.769226 6.22386 0.993084 6 1.26923 6H7.16153C7.43768 6 7.66153 6.22386 7.66153 6.5ZM4.23064 3.6945C3.68095 4.25278 2.78058 4.25278 2.23088 3.6945L1.60357 3.05738C1.41381 2.86466 1.41381 2.55534 1.60357 2.36262C1.79726 2.1659 2.11443 2.1656 2.3085 2.36194L2.59323 2.65C2.94392 3.0048 3.51705 3.00425 3.86705 2.64877L6.12011 0.360506C6.31505 0.162524 6.63452 0.163088 6.82876 0.361758C7.01783 0.55515 7.01729 0.864315 6.82753 1.05704L4.23064 3.6945Z"
-      fill="#ABBFF0"
-    />
-  </svg>
-);
-
-export const DotDotDotIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="7"
-    height="2"
-    viewBox="0 0 7 2"
-    fill="none"
-  >
-    <path
-      d="M5.92308 1.5C5.51523 1.5 5.18461 1.16421 5.18461 0.75C5.18461 0.335786 5.51523 0 5.92308 0C6.33092 0 6.66154 0.335786 6.66154 0.75C6.66154 0.948912 6.58374 1.13968 6.44525 1.28033C6.30676 1.42098 6.11893 1.5 5.92308 1.5ZM3.70769 1.5C3.29985 1.5 2.96923 1.16421 2.96923 0.75C2.96923 0.335786 3.29985 0 3.70769 0C4.11553 0 4.44615 0.335786 4.44615 0.75C4.44615 0.948912 4.36835 1.13968 4.22986 1.28033C4.09137 1.42098 3.90354 1.5 3.70769 1.5ZM1.49231 1.5C1.08447 1.5 0.753845 1.16421 0.753845 0.75C0.753845 0.335786 1.08447 0 1.49231 0C1.90015 0 2.23077 0.335786 2.23077 0.75C2.23077 0.948912 2.15297 1.13968 2.01448 1.28033C1.87599 1.42098 1.68816 1.5 1.49231 1.5Z"
-      fill="#DDE0E5"
-    />
-  </svg>
-);
-
-export const OperationalIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="5"
-    height="6"
-    viewBox="0 0 5 6"
-    fill="none"
-  >
-    <ellipse cx="2.5" cy="3.3636" rx="2.5" ry="2.45455" fill="#1BA546" />
-  </svg>
-);
-
-export const InProgressIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="5"
-    height="6"
-    viewBox="0 0 5 6"
-    fill="none"
-  >
-    <circle cx="2.5" cy="3" r="2.5" fill="#D6D6D6" />
-  </svg>
-);
-
-export const SmallSizeIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="6"
-    height="9"
-    viewBox="0 0 6 9"
-    fill="none"
-  >
-    <path
-      d="M0.590391 4.57726L4.22789 0.45262L3.63751 3.54051L5.60318 3.90278L1.96568 8.02742L2.55606 4.93954L0.590391 4.57726Z"
-      stroke="#2E3A59"
-      strokeOpacity="0.5"
-      strokeWidth="0.75"
-      strokeLinejoin="round"
-    />
-  </svg>
-);
-
-export const SmallLandBasedWindIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="6"
-    height="9"
-    viewBox="0 0 6 9"
-    fill="none"
-  >
-    <path
-      d="M5.91913 5.83222L5.28285 4.65825C5.20579 4.52274 5.07931 4.42411 4.93121 4.38404L4.18609 4.19266C4.17493 4.05841 4.14144 3.92988 4.07725 3.81276L5.18238 1.93042C5.36378 1.62193 5.2661 1.22204 4.9647 1.03638C4.84573 0.964435 4.70718 0.9339 4.56983 0.949349C4.43247 0.964799 4.3037 1.0254 4.20283 1.12207L3.24562 2.03896C3.13399 2.14465 3.0698 2.29603 3.0698 2.45313V3.4243C2.93864 3.47286 2.82142 3.55283 2.72654 3.65852H0.636286C0.279073 3.65852 0 3.94416 0 4.30977C0 4.60112 0.186979 4.85534 0.46047 4.93532L1.71909 5.30379C1.76932 5.31807 1.82234 5.32664 1.87258 5.32664C1.97304 5.32664 2.07351 5.29807 2.16002 5.2438L2.79073 4.86105C2.86608 4.93532 2.96096 4.9953 3.0698 5.03529V7.94307C2.76282 7.94307 2.51165 8.48578 2.51165 8.79998H4.18609C4.18609 8.48578 3.93493 7.94307 3.62795 7.94307V5.29236L4.91447 6.60915C5.04005 6.73768 5.20191 6.80052 5.36378 6.80052C5.52564 6.80052 5.69029 6.73768 5.81308 6.60915C6.01402 6.40349 6.05867 6.08643 5.91913 5.83222ZM1.87537 4.75537L0.61396 4.38689C0.580471 4.37261 0.558145 4.34405 0.558145 4.30977C0.558145 4.26693 0.594425 4.22979 0.636286 4.22979H2.51165C2.51165 4.27264 2.51165 4.31548 2.52282 4.35547L1.87537 4.75537ZM3.34887 4.51543C3.19538 4.51543 3.0698 4.38689 3.0698 4.22979C3.0698 4.07269 3.19538 3.94416 3.34887 3.94416C3.50236 3.94416 3.62795 4.07269 3.62795 4.22979C3.62795 4.38689 3.50236 4.51543 3.34887 4.51543ZM3.62795 3.41287V2.45313L4.58516 1.53624C4.6047 1.51625 4.63819 1.50196 4.68005 1.52481C4.71633 1.54767 4.73028 1.59622 4.70796 1.63336L3.66422 3.41573L3.62795 3.41287ZM5.41959 6.2064C5.40564 6.22068 5.35261 6.2521 5.30238 6.2064L3.95446 4.82106C3.97679 4.80107 3.99911 4.76965 4.02144 4.74108L4.79726 4.94389L5.43355 6.10928C5.45029 6.1407 5.44192 6.18069 5.41959 6.2064Z"
-      fill="#96CEE5"
-    />
-  </svg>
-);
-
-export const SmallSolarPowerIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="10"
-    height="9"
-    viewBox="0 0 10 9"
-    fill="none"
-  >
-    <path
-      d="M4.74286 1.05V0.45C4.74286 0.330653 4.79253 0.216193 4.88094 0.131802C4.96935 0.0474106 5.08926 0 5.21429 0C5.33932 0 5.45923 0.0474106 5.54764 0.131802C5.63605 0.216193 5.68571 0.330653 5.68571 0.45V1.05C5.68571 1.16935 5.63605 1.28381 5.54764 1.3682C5.45923 1.45259 5.33932 1.5 5.21429 1.5C5.08926 1.5 4.96935 1.45259 4.88094 1.3682C4.79253 1.28381 4.74286 1.16935 4.74286 1.05ZM7.88571 4.5C7.88571 5.00434 7.72904 5.49736 7.4355 5.9167C7.14196 6.33605 6.72474 6.66289 6.2366 6.85589C5.74846 7.0489 5.21132 7.0994 4.69312 7.001C4.17491 6.90261 3.69891 6.65975 3.3253 6.30312C2.95169 5.9465 2.69727 5.49213 2.59419 4.99748C2.49111 4.50283 2.54401 3.99011 2.74621 3.52416C2.9484 3.05821 3.29081 2.65995 3.73012 2.37975C4.16943 2.09956 4.68593 1.95 5.21429 1.95C5.92257 1.95069 6.60163 2.21958 7.10247 2.69765C7.6033 3.17571 7.88499 3.82391 7.88571 4.5ZM6.94286 4.5C6.94286 4.17366 6.84148 3.85465 6.65154 3.58331C6.4616 3.31197 6.19164 3.10048 5.87578 2.9756C5.55993 2.85071 5.21237 2.81804 4.87706 2.8817C4.54175 2.94537 4.23375 3.10252 3.992 3.33327C3.75026 3.56403 3.58563 3.85803 3.51893 4.1781C3.45223 4.49817 3.48646 4.82993 3.61729 5.13143C3.74813 5.43293 3.96968 5.69062 4.25394 5.87192C4.5382 6.05323 4.87241 6.15 5.21429 6.15C5.67257 6.1495 6.11194 5.97551 6.436 5.66618C6.76005 5.35685 6.94234 4.93746 6.94286 4.5ZM2.20932 2.26837C2.25317 2.31023 2.30523 2.34344 2.36253 2.36609C2.41982 2.38875 2.48123 2.40041 2.54325 2.40041C2.60527 2.40041 2.66668 2.38875 2.72397 2.36609C2.78127 2.34344 2.83333 2.31023 2.87718 2.26837C2.92103 2.22652 2.95582 2.17682 2.97955 2.12213C3.00328 2.06744 3.0155 2.00882 3.0155 1.94963C3.0155 1.89043 3.00328 1.83181 2.97955 1.77712C2.95582 1.72243 2.92103 1.67273 2.87718 1.63088L2.40575 1.18088C2.31719 1.09634 2.19707 1.04884 2.07182 1.04884C1.94657 1.04884 1.82646 1.09634 1.73789 1.18088C1.64933 1.26541 1.59958 1.38007 1.59958 1.49963C1.59958 1.61918 1.64933 1.73384 1.73789 1.81838L2.20932 2.26837ZM2.20932 6.73088L1.73789 7.18088C1.69404 7.22273 1.65926 7.27243 1.63552 7.32712C1.61179 7.38181 1.59958 7.44043 1.59958 7.49963C1.59958 7.55882 1.61179 7.61744 1.63552 7.67213C1.65926 7.72682 1.69404 7.77652 1.73789 7.81838C1.82646 7.90291 1.94657 7.95041 2.07182 7.95041C2.13384 7.95041 2.19525 7.93875 2.25254 7.91609C2.30984 7.89344 2.3619 7.86023 2.40575 7.81838L2.87718 7.36838C2.96574 7.28384 3.0155 7.16918 3.0155 7.04963C3.0155 6.93007 2.96574 6.81541 2.87718 6.73088C2.78862 6.64634 2.6685 6.59884 2.54325 6.59884C2.418 6.59884 2.29788 6.64634 2.20932 6.73088ZM7.88571 2.4C7.94764 2.40005 8.00897 2.38845 8.0662 2.36586C8.12343 2.34328 8.17544 2.31015 8.21925 2.26837L8.69068 1.81838C8.73453 1.77652 8.76932 1.72682 8.79305 1.67213C8.81678 1.61744 8.829 1.55882 8.829 1.49963C8.829 1.44043 8.81678 1.38181 8.79305 1.32712C8.76932 1.27243 8.73453 1.22273 8.69068 1.18088C8.64683 1.13902 8.59477 1.10581 8.53747 1.08316C8.48018 1.0605 8.41877 1.04884 8.35675 1.04884C8.29473 1.04884 8.23333 1.0605 8.17603 1.08316C8.11873 1.10581 8.06667 1.13902 8.02282 1.18088L7.55139 1.63088C7.48507 1.69381 7.43985 1.77416 7.4215 1.8617C7.40315 1.94924 7.41249 2.04001 7.44833 2.12246C7.48417 2.20492 7.5449 2.27534 7.62278 2.32476C7.70067 2.37418 7.79219 2.40037 7.88571 2.4ZM8.21925 6.73163C8.13069 6.64709 8.01057 6.59959 7.88532 6.59959C7.76007 6.59959 7.63996 6.64709 7.55139 6.73163C7.46283 6.81616 7.41307 6.93082 7.41307 7.05038C7.41307 7.16993 7.46283 7.28459 7.55139 7.36913L8.02282 7.81913C8.11138 7.90366 8.2315 7.95116 8.35675 7.95116C8.482 7.95116 8.60212 7.90366 8.69068 7.81913C8.77924 7.73459 8.829 7.61993 8.829 7.50037C8.829 7.38082 8.77924 7.26616 8.69068 7.18162L8.21925 6.73163ZM2.07143 4.5C2.07143 4.38065 2.02176 4.26619 1.93335 4.1818C1.84494 4.09741 1.72503 4.05 1.6 4.05H0.971429C0.846398 4.05 0.726488 4.09741 0.638078 4.1818C0.549668 4.26619 0.5 4.38065 0.5 4.5C0.5 4.61935 0.549668 4.73381 0.638078 4.8182C0.726488 4.90259 0.846398 4.95 0.971429 4.95H1.6C1.72503 4.95 1.84494 4.90259 1.93335 4.8182C2.02176 4.73381 2.07143 4.61935 2.07143 4.5ZM5.21429 7.5C5.08926 7.5 4.96935 7.54741 4.88094 7.6318C4.79253 7.71619 4.74286 7.83065 4.74286 7.95V8.55C4.74286 8.66935 4.79253 8.78381 4.88094 8.8682C4.96935 8.95259 5.08926 9 5.21429 9C5.33932 9 5.45923 8.95259 5.54764 8.8682C5.63605 8.78381 5.68571 8.66935 5.68571 8.55V7.95C5.68571 7.83065 5.63605 7.71619 5.54764 7.6318C5.45923 7.54741 5.33932 7.5 5.21429 7.5ZM9.45714 4.05H8.82857C8.70354 4.05 8.58363 4.09741 8.49522 4.1818C8.40681 4.26619 8.35714 4.38065 8.35714 4.5C8.35714 4.61935 8.40681 4.73381 8.49522 4.8182C8.58363 4.90259 8.70354 4.95 8.82857 4.95H9.45714C9.58217 4.95 9.70208 4.90259 9.79049 4.8182C9.8789 4.73381 9.92857 4.61935 9.92857 4.5C9.92857 4.38065 9.8789 4.26619 9.79049 4.1818C9.70208 4.09741 9.58217 4.05 9.45714 4.05Z"
-      fill="#FFBE54"
-    />
-  </svg>
-);
-
-export const SmallHydroelectricIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="12"
-    height="9"
-    viewBox="0 0 12 9"
-    fill="none"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M2.59162 0.919149C2.93395 0.359137 3.64985 0.380103 4.01153 0.863975C4.49872 1.51557 5.18536 2.15504 6.27187 2.15504C7.37723 2.15504 8.05345 1.64192 8.51385 0.998598C8.87156 0.498174 9.6346 0.417621 10.0097 1.01129C10.3163 1.49682 10.6517 1.88303 11.3135 2.0469C11.4075 2.07309 11.4886 2.13902 11.5395 2.23068C11.5905 2.32234 11.6073 2.43249 11.5863 2.5377C11.5654 2.64292 11.5084 2.7349 11.4274 2.79409C11.3464 2.85328 11.2479 2.875 11.1527 2.85464C10.2394 2.62953 9.76408 2.06235 9.40043 1.48688C9.37661 1.44881 9.33345 1.42068 9.26846 1.42233C9.23556 1.42281 9.20314 1.43119 9.17342 1.4469C9.14371 1.46262 9.1174 1.48529 9.0963 1.51337C8.51286 2.32883 7.63125 2.98264 6.27187 2.98264C4.87131 2.98264 3.99466 2.13738 3.439 1.39254C3.4232 1.37156 3.40336 1.35482 3.38094 1.34355C3.35852 1.33228 3.33411 1.32677 3.30951 1.32743C3.28943 1.32658 3.26948 1.3314 3.25153 1.34143C3.23357 1.35146 3.21819 1.36639 3.20681 1.38481C2.83671 1.99007 2.36191 2.61518 1.391 2.85464C1.34276 2.86808 1.29259 2.87063 1.24344 2.86214C1.1943 2.85365 1.14719 2.83429 1.10489 2.80521C1.06259 2.77613 1.02596 2.73791 0.997166 2.69281C0.968375 2.64772 0.948004 2.59666 0.937258 2.54266C0.926511 2.48866 0.925608 2.43281 0.934601 2.37842C0.943594 2.32402 0.9623 2.27219 0.989615 2.22597C1.01693 2.17975 1.0523 2.14009 1.09363 2.10933C1.13496 2.07858 1.18142 2.05735 1.23026 2.0469C1.93078 1.87365 2.26616 1.45102 2.59162 0.919149ZM2.59162 3.67783C2.93395 3.11782 3.64985 3.13878 4.01153 3.62265C4.49872 4.27425 5.18536 4.91372 6.27187 4.91372C7.37723 4.91372 8.05345 4.4006 8.51385 3.75728C8.87156 3.25685 9.6346 3.1763 10.0097 3.76997C10.3163 4.2555 10.6517 4.64171 11.3135 4.80558C11.4075 4.83177 11.4886 4.8977 11.5395 4.98936C11.5905 5.08102 11.6073 5.19117 11.5863 5.29638C11.5654 5.4016 11.5084 5.49358 11.4274 5.55277C11.3464 5.61196 11.2479 5.63368 11.1527 5.61332C10.2394 5.38821 9.76408 4.82102 9.40043 4.24556C9.37661 4.20749 9.33345 4.17936 9.26846 4.18101C9.23556 4.18149 9.20314 4.18987 9.17342 4.20558C9.14371 4.2213 9.1174 4.24397 9.0963 4.27205C8.51286 5.08806 7.63125 5.74132 6.27187 5.74132C4.87131 5.74132 3.99466 4.89551 3.439 4.15122C3.4232 4.13024 3.40336 4.1135 3.38094 4.10223C3.35852 4.09096 3.33411 4.08545 3.30951 4.08611C3.28943 4.08526 3.26948 4.09008 3.25153 4.10011C3.23357 4.11014 3.21819 4.12507 3.20681 4.14349C2.83671 4.74875 2.36191 5.37386 1.391 5.61332C1.29462 5.63702 1.19371 5.61718 1.11049 5.55815C1.02726 5.49912 0.968533 5.40574 0.947217 5.29855C0.925901 5.19137 0.943746 5.07915 0.996826 4.9866C1.04991 4.89404 1.13387 4.82873 1.23026 4.80502C1.93078 4.63288 2.26616 4.2097 2.59162 3.67783ZM2.59162 6.43651C2.93395 5.8765 3.64985 5.89746 4.01153 6.38133C4.49872 7.03293 5.18536 7.6724 6.27187 7.6724C7.37723 7.6724 8.05345 7.15928 8.51385 6.51596C8.87156 6.01553 9.6346 5.93498 10.0097 6.52865C10.3163 7.01417 10.6517 7.40039 11.3135 7.56426C11.4075 7.59045 11.4886 7.65638 11.5395 7.74804C11.5905 7.8397 11.6073 7.94985 11.5863 8.05506C11.5654 8.16028 11.5084 8.25226 11.4274 8.31145C11.3464 8.37064 11.2479 8.39236 11.1527 8.372C10.2394 8.14689 9.76408 7.5797 9.40043 7.00424C9.37661 6.96617 9.33345 6.93804 9.26846 6.93969C9.23556 6.94017 9.20314 6.94855 9.17342 6.96426C9.14371 6.97998 9.1174 7.00265 9.0963 7.03073C8.51286 7.84674 7.63125 8.5 6.27187 8.5C4.87131 8.5 3.99466 7.65474 3.439 6.9099C3.4232 6.88892 3.40336 6.87218 3.38094 6.86091C3.35852 6.84963 3.33411 6.84413 3.30951 6.84479C3.28943 6.84394 3.26948 6.84876 3.25153 6.85879C3.23357 6.86882 3.21819 6.88375 3.20681 6.90217C2.83671 7.50743 2.36191 8.13254 1.391 8.372C1.29462 8.3957 1.19371 8.37586 1.11049 8.31683C1.02726 8.2578 0.968533 8.16442 0.947217 8.05723C0.925901 7.95005 0.943746 7.83783 0.996826 7.74528C1.04991 7.65272 1.13387 7.58741 1.23026 7.5637C1.93078 7.39156 2.26616 6.96838 2.59162 6.43651Z"
-      fill="#0072BB"
-    />
-  </svg>
-);
-
-export const SmallOffshoreWindIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="9"
-    height="11"
-    viewBox="0 0 9 11"
-    fill="none"
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M1.86023 9.24538C2.12061 8.9049 2.66514 8.91764 2.94023 9.21183C3.3108 9.608 3.83306 9.99679 4.65948 9.99679C5.50023 9.99679 6.01457 9.68482 6.36476 9.29369C6.63683 8.98943 7.21721 8.94045 7.5025 9.3014C7.7357 9.5966 7.9908 9.83142 8.49419 9.93105C8.56569 9.94697 8.62737 9.98706 8.66613 10.0428C8.70489 10.0985 8.71766 10.1655 8.70174 10.2295C8.68581 10.2934 8.64244 10.3493 8.58085 10.3853C8.51926 10.4213 8.4443 10.4345 8.37193 10.4221C7.67721 10.2853 7.3157 9.94044 7.0391 9.59056C7.02099 9.56741 6.98816 9.55031 6.93872 9.55131C6.9137 9.5516 6.88904 9.5567 6.86644 9.56625C6.84384 9.57581 6.82383 9.58959 6.80778 9.60666C6.36401 10.1025 5.69344 10.5 4.65948 10.5C3.59419 10.5 2.9274 9.98606 2.50476 9.5332C2.49274 9.52045 2.47765 9.51027 2.4606 9.50341C2.44355 9.49656 2.42498 9.49321 2.40627 9.49362C2.39099 9.4931 2.37582 9.49603 2.36216 9.50213C2.34851 9.50823 2.33681 9.5173 2.32816 9.5285C2.04665 9.89649 1.68551 10.2766 0.947024 10.4221C0.910331 10.4303 0.872168 10.4319 0.83479 10.4267C0.797411 10.4215 0.761576 10.4098 0.729402 10.3921C0.697228 10.3744 0.669368 10.3512 0.647469 10.3238C0.625569 10.2963 0.610075 10.2653 0.601901 10.2325C0.593727 10.1996 0.59304 10.1657 0.59988 10.1326C0.606721 10.0995 0.620949 10.068 0.641725 10.0399C0.662501 10.0118 0.689403 9.98771 0.720841 9.96901C0.752278 9.95031 0.787614 9.9374 0.82476 9.93105C1.35759 9.82571 1.61268 9.56876 1.86023 9.24538Z"
-      fill="#4974E0"
-    />
-    <path
-      d="M2.50476 9.5332C2.9274 9.98606 3.59419 10.5 4.65948 10.5C5.69344 10.5 6.364 10.1025 6.80778 9.60666C6.82383 9.58959 6.84384 9.57581 6.86644 9.56625C6.88904 9.5567 6.9137 9.5516 6.93872 9.55131C6.98816 9.55031 7.02099 9.56741 7.0391 9.59056C7.3157 9.94044 7.67721 10.2853 8.37193 10.4221C8.4443 10.4345 8.51926 10.4213 8.58085 10.3853C8.64244 10.3493 8.68581 10.2934 8.70174 10.2295C8.71766 10.1655 8.70489 10.0985 8.66613 10.0428C8.62737 9.98706 8.56569 9.94697 8.49419 9.93105C7.9908 9.83142 7.7357 9.5966 7.5025 9.3014C7.21721 8.94045 6.63683 8.98943 6.36476 9.29369C6.01457 9.68482 5.50023 9.99679 4.65948 9.99679C3.83306 9.99679 3.3108 9.608 2.94023 9.21183C2.66514 8.91764 2.12061 8.9049 1.86023 9.24538C1.61268 9.56876 1.35759 9.82571 0.82476 9.93105C0.787614 9.9374 0.752278 9.95031 0.720841 9.96901C0.689403 9.98771 0.662501 10.0118 0.641725 10.0399C0.620949 10.068 0.606721 10.0995 0.59988 10.1326C0.59304 10.1657 0.593727 10.1996 0.601901 10.2325C0.610075 10.2653 0.625569 10.2963 0.647469 10.3238C0.669368 10.3512 0.697228 10.3744 0.729402 10.3921C0.761576 10.4098 0.797411 10.4215 0.83479 10.4267C0.872168 10.4319 0.910331 10.4303 0.947024 10.4221C1.68551 10.2766 2.04665 9.89649 2.32816 9.5285C2.33681 9.5173 2.34851 9.50823 2.36216 9.50213C2.37582 9.49603 2.39099 9.4931 2.40627 9.49362C2.42498 9.49321 2.44355 9.49656 2.4606 9.50341C2.47765 9.51027 2.49274 9.52045 2.50476 9.5332ZM2.50476 9.5332L2.7225 9.37218"
-      stroke="#2E3A59"
-      strokeWidth="0.1"
-    />
-    <path
-      d="M8.04066 6.85855L7.2403 5.33102C7.14338 5.1547 6.98428 5.02636 6.798 4.97422L5.86074 4.7252C5.8467 4.55052 5.80457 4.38328 5.72383 4.23089L7.11393 1.78164C7.3421 1.38024 7.21924 0.859911 6.84012 0.618331C6.69047 0.524723 6.5162 0.484992 6.34342 0.505094C6.17065 0.525196 6.00868 0.604048 5.8818 0.729829L4.67775 1.92287C4.53734 2.06038 4.4566 2.25736 4.4566 2.46178V3.72543C4.29161 3.78861 4.14418 3.89268 4.02483 4.0302H1.39557C0.94625 4.0302 0.595215 4.40186 0.595215 4.87759C0.595215 5.25668 0.830408 5.58746 1.17442 5.69153L2.75759 6.17097C2.82078 6.18956 2.88747 6.20071 2.95066 6.20071C3.07703 6.20071 3.2034 6.16354 3.31222 6.09292L4.10556 5.5949C4.20034 5.69153 4.31969 5.76958 4.4566 5.82161V9.9825C4.07046 9.9825 3.75453 9.90249 3.75453 10.3113H5.86074C5.86074 9.90249 5.54481 9.9825 5.15867 9.9825V6.15611L6.77694 7.86947C6.9349 8.03672 7.1385 8.11849 7.3421 8.11849C7.5457 8.11849 7.75282 8.03672 7.90727 7.86947C8.16002 7.60188 8.21618 7.18933 8.04066 6.85855ZM2.95417 5.45738L1.36749 4.97794C1.32537 4.95935 1.29728 4.92219 1.29728 4.87759C1.29728 4.82184 1.34292 4.77352 1.39557 4.77352H3.75453C3.75453 4.82927 3.75453 4.88502 3.76857 4.93705L2.95417 5.45738ZM4.80763 5.14518C4.61456 5.14518 4.4566 4.97794 4.4566 4.77352C4.4566 4.56911 4.61456 4.40186 4.80763 4.40186C5.0007 4.40186 5.15867 4.56911 5.15867 4.77352C5.15867 4.97794 5.0007 5.14518 4.80763 5.14518ZM5.15867 3.71057V2.46178L6.36272 1.26874C6.38729 1.24272 6.42941 1.22414 6.48207 1.25387C6.5277 1.28361 6.54526 1.34679 6.51717 1.39511L5.2043 3.71428L5.15867 3.71057ZM7.41231 7.34543C7.39476 7.36401 7.32806 7.40489 7.26488 7.34543L5.56938 5.54286C5.59746 5.51685 5.62554 5.47596 5.65363 5.4388L6.6295 5.70268L7.42986 7.21906C7.45093 7.25995 7.44039 7.31198 7.41231 7.34543Z"
-      fill="#4974E0"
-    />
-  </svg>
-);
-
-export const SmallGeothermalIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="12"
-    height="9"
-    viewBox="0 0 12 9"
-    fill="none"
-  >
-    <path
-      d="M5.4716 2.13159L5.3978 2.11821C5.38121 2.20972 5.33022 2.29396 5.25235 2.35538C5.17436 2.41689 5.07475 2.45123 4.97109 2.45132L4.97087 2.45132C4.94663 2.45142 4.92244 2.4495 4.89859 2.4456L4.89806 2.44551C4.84129 2.43664 4.78717 2.41746 4.73879 2.38925C4.69041 2.36103 4.64883 2.32441 4.6162 2.28171C4.58359 2.23902 4.56054 2.19105 4.5481 2.1406C4.53567 2.09017 4.53403 2.03802 4.54325 1.98704L4.54328 1.98684C4.6383 1.45307 4.93408 0.967066 5.37924 0.616347C5.82452 0.265543 6.38992 0.0733078 6.97441 0.0750115H6.97462C7.09111 0.0750115 7.20186 0.118203 7.28274 0.193607C7.36344 0.268846 7.40769 0.369721 7.40769 0.473695C7.40769 0.577669 7.36344 0.678545 7.28274 0.753783C7.20189 0.82916 7.09119 0.872348 6.97475 0.872379M5.4716 2.13159L5.39775 2.11849C5.45994 1.76791 5.65377 1.45058 5.94298 1.22272C6.23205 0.994968 6.59782 0.871128 6.97475 0.872379M5.4716 2.13159C5.53044 1.79992 5.71399 1.49862 5.98939 1.28163C6.2648 1.06465 6.61406 0.946155 6.97462 0.947379M5.4716 2.13159C5.4516 2.24188 5.39039 2.34202 5.29879 2.41427C5.2072 2.48652 5.09113 2.52621 4.97116 2.52632C4.94277 2.52643 4.91444 2.52419 4.88648 2.51961C4.82063 2.50933 4.75761 2.48704 4.70101 2.45403C4.6444 2.42102 4.59534 2.37794 4.5566 2.32724C4.51787 2.27654 4.49024 2.21922 4.47528 2.15855C4.46033 2.09789 4.45834 2.03507 4.46944 1.97369L6.97462 1.1542e-05C7.10937 1.1542e-05 7.2386 0.0499174 7.33388 0.13875C7.42916 0.227583 7.48269 0.348066 7.48269 0.473695C7.48269 0.599324 7.42916 0.719807 7.33388 0.80864C7.2386 0.897473 7.10937 0.947379 6.97462 0.947379M6.97475 0.872379C6.97479 0.872379 6.97484 0.872379 6.97488 0.872379L6.97462 0.947379M6.97475 0.872379C6.97471 0.872379 6.97467 0.872379 6.97462 0.872379V0.947379M1.90033 8.12763H1.94651L1.9673 8.0864C2.49761 7.03451 3.08394 5.52825 3.1489 3.91843L3.14891 3.91823C3.15581 3.7354 3.23883 3.56107 3.38225 3.43264C3.52583 3.30407 3.71848 3.23194 3.91953 3.2329H3.91989H7.66009H7.66016C7.86041 3.2327 8.05215 3.30456 8.19564 3.43219C8.33896 3.55966 8.42294 3.73265 8.43195 3.9147C8.51755 5.67518 9.39173 7.36037 9.82162 8.09068L9.84338 8.12763H9.88626H10.7004C10.8169 8.12763 10.9277 8.17082 11.0085 8.24623C11.0892 8.32147 11.1335 8.42234 11.1335 8.52632C11.1335 8.63029 11.0892 8.73117 11.0085 8.80641C10.9277 8.88181 10.8169 8.925 10.7004 8.925H1.21656C1.10008 8.925 0.989326 8.88181 0.908449 8.80641C0.82775 8.73117 0.783496 8.63029 0.783496 8.52632C0.783496 8.42234 0.827749 8.32147 0.908449 8.24623C0.989326 8.17082 1.10008 8.12763 1.21656 8.12763H1.90033ZM7.57444 4.10001L7.56954 4.03027H7.49962H6.82221H6.74212L6.74737 4.11018C6.84821 5.64749 7.40875 7.07643 7.91829 8.08642L7.93908 8.12763H7.98525H8.73803H8.86331L8.80414 8.01721C8.30968 7.09442 7.68345 5.6503 7.57444 4.10001ZM2.94711 8.021L2.89749 8.12763H3.01511H6.86793H6.98555L6.93593 8.021C6.44458 6.96492 5.96361 5.58223 5.87629 4.10086L5.87213 4.03027H5.80142H4.08162H4.01091L4.00675 4.10086C3.91944 5.58223 3.43846 6.96492 2.94711 8.021ZM6.49716 1.77255C6.57804 1.69715 6.68878 1.65396 6.80527 1.65396H6.97462C7.3073 1.65396 7.62731 1.5308 7.86403 1.31009C8.10094 1.08922 8.23511 0.788483 8.23511 0.473695C8.23511 0.369721 8.27936 0.268846 8.36006 0.193607C8.44094 0.118203 8.55169 0.0750115 8.66817 0.0750115C8.78466 0.0750115 8.89541 0.118203 8.97628 0.193607C9.05696 0.268823 9.10121 0.369661 9.10124 0.473603C9.10059 0.99618 8.87769 1.49834 8.47951 1.86957C8.08116 2.24097 7.54001 2.45072 6.97454 2.45132H6.80527C6.68878 2.45132 6.57804 2.40813 6.49716 2.33273C6.41646 2.25749 6.37221 2.15662 6.37221 2.05264C6.37221 1.94867 6.41646 1.84779 6.49716 1.77255Z"
-      fill="#4D8B31"
-      stroke="white"
-      strokeWidth="0.15"
-    />
-  </svg>
-);
-
-export const SmallEnergyStorageIcon = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="13"
-    height="9"
-    viewBox="0 0 13 9"
-    fill="none"
-  >
-    <path
-      d="M2.87528 6.40012C2.725 6.40012 2.58086 6.34042 2.47459 6.23415C2.36832 6.12788 2.30862 5.98374 2.30862 5.83346V3.16679C2.30862 3.0165 2.36832 2.87237 2.47459 2.7661C2.58086 2.65982 2.725 2.60012 2.87528 2.60012H6.77529V6.40012H2.87528Z"
-      fill="#4896BC"
-      stroke="white"
-      strokeWidth="0.2"
-    />
-    <path
-      d="M11.9228 6.4H11.8228V6.5C11.8228 7.00391 11.6226 7.48718 11.2663 7.8435C10.91 8.19982 10.4267 8.4 9.92278 8.4H2.2085C1.70458 8.4 1.22131 8.19982 0.864993 7.8435C0.508674 7.48718 0.308496 7.00391 0.308496 6.5V2.5C0.308496 1.99609 0.508674 1.51282 0.864993 1.1565C1.22131 0.800178 1.70458 0.6 2.2085 0.6H9.92278C10.4267 0.6 10.91 0.800178 11.2663 1.1565C11.6226 1.51282 11.8228 1.99609 11.8228 2.5V2.6H11.9228C12.0731 2.6 12.2172 2.6597 12.3235 2.76597C12.4297 2.87224 12.4894 3.01638 12.4894 3.16667V5.83333C12.4894 5.98362 12.4297 6.12776 12.3235 6.23403C12.2172 6.3403 12.0731 6.4 11.9228 6.4ZM9.92278 1.73333H2.2085C2.00516 1.73333 1.81016 1.81411 1.66638 1.95789C1.5226 2.10166 1.44183 2.29667 1.44183 2.5V6.5C1.44183 6.70333 1.5226 6.89834 1.66638 7.04212C1.81016 7.18589 2.00516 7.26667 2.2085 7.26667H9.92278C10.1261 7.26667 10.3211 7.18589 10.4649 7.04212C10.6087 6.89834 10.6894 6.70333 10.6894 6.5V2.5C10.6894 2.29667 10.6087 2.10166 10.4649 1.95788C10.3211 1.81411 10.1261 1.73333 9.92278 1.73333Z"
-      fill="#4896BC"
-      stroke="white"
-      strokeWidth="0.2"
-    />
-  </svg>
-);
-
-export const SmallPumpedStorage = () => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    width="12"
-    height="9"
-    viewBox="0 0 12 9"
-    fill="none"
-  >
-    <path
-      d="M9.76821 7.3395V8.1C9.76821 8.33869 9.65444 8.56761 9.45192 8.7364C9.24941 8.90518 8.97474 9 8.68834 9H3.28902C3.00262 9 2.72795 8.90518 2.52544 8.7364C2.32292 8.56761 2.20915 8.33869 2.20915 8.1V6.435C1.96078 6.354 1.64222 6.3 1.12929 6.3C0.986089 6.3 0.848755 6.25259 0.747498 6.1682C0.646241 6.08381 0.589355 5.96935 0.589355 5.85C0.589355 5.73065 0.646241 5.61619 0.747498 5.5318C0.848755 5.44741 0.986089 5.4 1.12929 5.4C1.57203 5.4 1.92299 5.436 2.20915 5.4945V4.635C1.96078 4.554 1.64222 4.5 1.12929 4.5C0.986089 4.5 0.848755 4.45259 0.747498 4.3682C0.646241 4.28381 0.589355 4.16935 0.589355 4.05C0.589355 3.93065 0.646241 3.81619 0.747498 3.7318C0.848755 3.64741 0.986089 3.6 1.12929 3.6C1.57203 3.6 1.92299 3.636 2.20915 3.6945V2.835C1.96078 2.754 1.64222 2.7 1.12929 2.7C0.986089 2.7 0.848755 2.65259 0.747498 2.5682C0.646241 2.48381 0.589355 2.36935 0.589355 2.25C0.589355 2.13065 0.646241 2.01619 0.747498 1.9318C0.848755 1.84741 0.986089 1.8 1.12929 1.8C1.57203 1.8 1.92299 1.836 2.20915 1.8945V0.9C2.20915 0.661305 2.32292 0.432387 2.52544 0.263604C2.72795 0.0948211 3.00262 0 3.28902 0H8.68834C8.97474 0 9.24941 0.0948211 9.45192 0.263604C9.65444 0.432387 9.76821 0.661305 9.76821 0.9V1.872C10.7293 2.0115 11.1396 2.3085 11.2314 2.3805C11.442 2.556 11.442 2.844 11.2314 3.0195C11.0209 3.195 10.7401 3.1905 10.4647 3.0195C10.4647 3.0195 9.90319 2.7 8.68834 2.7C8.00803 2.7 7.55988 2.8845 7.04155 3.105C6.48002 3.3345 5.8483 3.6 4.90882 3.6C4.71444 3.6 4.53626 3.6 4.36888 3.582V2.6775C4.53086 2.7 4.70904 2.7 4.90882 2.7C5.58913 2.7 6.03727 2.5155 6.55561 2.2995C7.11714 2.0655 7.75426 1.8 8.68834 1.8V0.9H3.28902V8.1H8.68834V7.2C9.49824 7.2 9.75201 7.3305 9.76821 7.3395ZM8.68834 3.6C7.75426 3.6 7.11714 3.8655 6.55561 4.0995C6.03727 4.3155 5.58913 4.5 4.90882 4.5C4.70904 4.5 4.53086 4.5 4.36888 4.4775V5.382C4.53626 5.4 4.71444 5.4 4.90882 5.4C5.8483 5.4 6.48002 5.1345 7.04155 4.905C7.55988 4.689 8.00803 4.5 8.68834 4.5C9.90319 4.5 10.4647 4.8195 10.4647 4.8195C10.7401 4.995 11.0209 4.995 11.2314 4.8195C11.442 4.644 11.442 4.3605 11.2314 4.1805C11.118 4.086 10.4431 3.6 8.68834 3.6ZM8.68834 5.4C7.75426 5.4 7.11714 5.6655 6.55561 5.8995C6.03727 6.1155 5.58913 6.3 4.90882 6.3C4.70904 6.3 4.53086 6.3 4.36888 6.2775V7.182C4.53626 7.2 4.71444 7.2 4.90882 7.2C5.8483 7.2 6.48002 6.9345 7.04155 6.705C7.55988 6.489 8.00803 6.3 8.68834 6.3C9.90319 6.3 10.4647 6.6195 10.4647 6.6195C10.7401 6.795 11.0209 6.795 11.2314 6.6195C11.442 6.444 11.442 6.1605 11.2314 5.9805C11.118 5.886 10.4431 5.4 8.68834 5.4Z"
-      fill="#68C6E2"
-    />
-  </svg>
-);
