+++ assets/KDM-Icons/icons.tsx
@@ -1,0 +1,29 @@
+export const CheckmarkIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="8"
+    height="7"
+    viewBox="0 0 8 7"
+    fill="none"
+  >
+    <path
+      d="M7.66153 6.5C7.66153 6.77614 7.43768 7 7.16153 7H1.26923C0.993084 7 0.769226 6.77614 0.769226 6.5C0.769226 6.22386 0.993084 6 1.26923 6H7.16153C7.43768 6 7.66153 6.22386 7.66153 6.5ZM4.23064 3.6945C3.68095 4.25278 2.78058 4.25278 2.23088 3.6945L1.60357 3.05738C1.41381 2.86466 1.41381 2.55534 1.60357 2.36262C1.79726 2.1659 2.11443 2.1656 2.3085 2.36194L2.59323 2.65C2.94392 3.0048 3.51705 3.00425 3.86705 2.64877L6.12011 0.360506C6.31505 0.162524 6.63452 0.163088 6.82876 0.361758C7.01783 0.55515 7.01729 0.864315 6.82753 1.05704L4.23064 3.6945Z"
+      fill="#ABBFF0"
+    />
+  </svg>
+);
+
+export const DotDotDotIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="7"
+    height="2"
+    viewBox="0 0 7 2"
+    fill="none"
+  >
+    <path
+      d="M5.92308 1.5C5.51523 1.5 5.18461 1.16421 5.18461 0.75C5.18461 0.335786 5.51523 0 5.92308 0C6.33092 0 6.66154 0.335786 6.66154 0.75C6.66154 0.948912 6.58374 1.13968 6.44525 1.28033C6.30676 1.42098 6.11893 1.5 5.92308 1.5ZM3.70769 1.5C3.29985 1.5 2.96923 1.16421 2.96923 0.75C2.96923 0.335786 3.29985 0 3.70769 0C4.11553 0 4.44615 0.335786 4.44615 0.75C4.44615 0.948912 4.36835 1.13968 4.22986 1.28033C4.09137 1.42098 3.90354 1.5 3.70769 1.5ZM1.49231 1.5C1.08447 1.5 0.753845 1.16421 0.753845 0.75C0.753845 0.335786 1.08447 0 1.49231 0C1.90015 0 2.23077 0.335786 2.23077 0.75C2.23077 0.948912 2.15297 1.13968 2.01448 1.28033C1.87599 1.42098 1.68816 1.5 1.49231 1.5Z"
+      fill="#DDE0E5"
+    />
+  </svg>
+);
